Fix implicit conversion from size_t to int warnings in Deiconified()
xconsole.c:423:14: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'int' [-Wshorten-64-to-32]
    oldlen = strlen (oldName);
           ~ ^~~~~~~~~~~~~~~~

xconsole.c:426:27: warning: implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
        newName = malloc (oldlen - 1);
                  ~~~~~~  ~~~~~~~^~~

xconsole.c:429:36: warning: implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
        strncpy (newName, oldName, oldlen - 2);
        ~~~~~~~                    ~~~~~~~^~~

+++ xconsole.c
@@ -413,7 +413,7 @@ Deiconified(Widget widget, XEvent *event, String *params, Cardinal *num_params)
     Arg	    arglist[1];
     char    *oldName;
     char    *newName;
-    int	    oldlen;
+    size_t  oldlen;
 
     iconified = False;
     if (!app_resources.notify || !notified)
